chore: README updated with accurate bikeshare info
The README.md file was updated to be consistent with the project flow
during the python portion of the Udacity nanodegree in data science.

The README.md file now contains the date of creation, a project description
and acknowledgements of various repositories and sources that were used
during the project.